Change `/logout` route to `POST`

+++ apps/web/app/[language]/logout/route.ts
@@ -1,23 +1,29 @@
-import { NextRequest } from 'next/server';
 import { db } from '@/lib/prisma';
 import { authCookieSettings, SessionCookieName } from '@/lib/auth/cookie';
 import { redirect } from 'next/navigation';
 import { cookies } from 'next/headers';
 
-export async function GET(request: NextRequest) {
-  // get session from cookie
-  // TODO: why not use `x-gw2t-session` header?
-  const sessionId = request.cookies.get(SessionCookieName)?.value ?? '';
+export async function POST() {
+  const cookieStore = await cookies();
 
+  if(!cookieStore.has(SessionCookieName)) {
+    // if there is no session cookie, redirect user to login page
+    redirect('/login');
+  }
+
+  const sessionId = cookieStore.get(SessionCookieName)!.value;
+
+  // try to delete session in db
   if(sessionId) {
-    // try to delete session in db
     await db.userSession.deleteMany({ where: { id: sessionId }});
   }
 
   // delete cookie
-  (await cookies()).delete(authCookieSettings);
+  cookieStore.delete(authCookieSettings);
+
+  // set logout cookie to show message
+  cookieStore.set('logout', '1', { maxAge: 2, httpOnly: true, path: '/login', secure: true });
 
   // redirect user to login page
-  // TODO: replace `?logout` with cookie to show message?
-  return redirect('/login?logout');
+  return redirect('/login');
 }
